+++ model/src/main/kotlin/io/github/openminigameserver/replay/model/recordable/RecordableVector.kt
@@ -1,3 +1,3 @@
 package io.github.openminigameserver.replay.model.recordable
 
-data class RecordableVector(val x: Float = 0f, val y: Float = 0f, val z: Float = 0f)
+data class RecordableVector(val x: Double = 0.0, val y: Double = 0.0, val z: Double = 0.0)
